requirements.txt
asgiref (3.2.3)
astroid (2.3.3)
autopep8 (1.4.4)
beautifulsoup4 (4.8.2)
certifi (2019.11.28)
cffi (1.13.2)
chardet (3.0.4)
cryptography (2.8)
Django (3.0.1)
django-crontab (0.7.1)
django-rest-knox (4.1.0)
django-webpack-loader (0.7.0)
djangorestframework (3.11.0)
idna (2.9)
isort (4.3.21)
lazy-object-proxy (1.4.3)
mccabe (0.6.1)
pip (9.0.1)
pkg-resources (0.0.0)
praw (6.5.1)
prawcore (1.0.1)
pycodestyle (2.5.0)
pycparser (2.19)
pylint (2.4.4)
pylint-protobuf (0.12)
python-dotenv (0.12.0)
pytz (2019.3)
requests (2.23.0)
setuptools (39.0.1)
six (1.13.0)
soupsieve (2.0)
sqlparse (0.3.0)
typed-ast (1.4.1)
update-checker (0.16)
urllib3 (1.25.8)
websocket-client (0.57.0)
wrapt (1.11.2)
